Add shared memory version
There is now a version of the density calculator that uses shared memory to accumulate densities and do a reduction to reduce the number of atomic adds to global memory.

NOTE: This version requires that there are as many particles as there are threads in a block, and there are still some differences between the output of the two methods.

Now the density memory is initialized to zero before accumulating to it. It worked before but now it is more robust.

Fixed vectors too large.

Re-introduced storing density to compare results of atomic and shared implementations.

Also removed the old main file since there is so much duplication now.