Step 1of 4 implemented
The first of the four steps on the whiteboard implemented. This is a qualite regression on the path to a more general method. This only works if n %2 == 0, and if gridDim.x % cell_particle_count == 0, and possibly more restrictions. Basically no cells can have multiple blocks work on them, and all particles must fill upp full blocks.