Randomize particle indices
Even if the particles are generated in an orderly fashion, their indices are not. This means plainly iterating over the positions will not yield consecutive cells.

This forces sorting the particles by cell for the shared method to have a chance of being effective. The sorting is performed currently but the result ins not used yet.