AoS instead of SoA
Switch from using two separate arrays for x and y and instead use an array of {x, y} structs. This is only used for the shared version, since the atomic one woul'nt benefit. The shared benefit because there is so much uncoalesced global access, so hitting both x and y while looking for x is beneficial.